refactor buttons to use internal buttonBase component
